backend/requirements.txt
fastapi
uvicorn
easyocr
opencv-python
pillow
spacy
python-multipart
